Refactor PackagesPage to use react-icons, improve link prefetching, and update icon components
Refactor IconsDocsPage to utilize dynamic loading for content sections and remove unnecessary state management

Enhance AfricanCityPage with dynamic import for GridAirQualityMonitor to improve loading performance

Update NetworkCoveragePageApi to replace AqLoading02 with FiLoader for a consistent icon usage

Modify NetworkCoverageHeader to replace AqDownload01 with FiDownload for improved icon consistency

Replace AqMarkerPin01 and AqSearchRefraction with FiMapPin and FiSearch in NetworkCoverageSidebar for uniformity

Add HomeDeferredSections component to dynamically load sections for improved performance

Create IconsPackageBrowser for searchable icon catalog with dynamic loading and improved user experience

Implement IconsPackageBrowserLoader to manage loading state for the icon browser

Create IconsDocsContent and IconsDocsContentLoader for structured documentation loading and improved user experience

Add NetworkCoverageDeferred component for dynamic loading of the NetworkCoveragePageApi
